--- OrientacaObjetos/src/EmailEnvio/UsanoBibliotecaDeEmail.java
@@ -1,15 +1,0 @@
-//package EmailEnvio;
-//
-//
-//
-//public class UsanoBibliotecaDeEmail {
-//
-//	public static void main(String[] args){
-//		
-//	Carteiro.enviar("jhonatandossantos.inter@outlook.com", "Teste envio de email", "FUNCIONOOU");
-//	
-//	System.out.println("deu");
-//
-//	}
-//
-//}
