--- OrientacaObjetos/src/LeituraArquivos/LeituraArquivos.java
@@ -1,16 +1,0 @@
-
-//
-//package LeituraArquivos;
-//
-//import java.io.IOException;
-//import java.nio.file.Files;
-//import java.nio.file.Path;
-//import java.nio.file.Paths;
-//import java.util.List;
-//
-//public class LeituraArquivos {
-//
-//	
-//		
-//	}
-
